# Conflicts:
#	dubbo-common/src/main/java/org/apache/dubbo/config/nested/TripleConfig.java
#	dubbo-rpc/dubbo-rpc-triple/src/main/java/org/apache/dubbo/rpc/protocol/tri/TripleProtocol.java

+++ dubbo-plugin/dubbo-triple-servlet/src/main/java/org/apache/dubbo/rpc/protocol/tri/servlet/HttpMetadataAdapter.java
@@ -72,8 +72,8 @@ final class HttpMetadataAdapter implements Http2Header {
     }
 
     @Override
-    public int id() {
-        return -1;
+    public long id() {
+        return -1L;
     }
 
     @Override
